refactor: extract blueprints into separate dictionary
